refactor: #14 controller 및 DTO폴더 구조 변경
