+++ src/main/java/LogITBackend/LogIT/service/UserCommandService.java
@@ -1,6 +1,6 @@
 package LogITBackend.LogIT.service;
 
-import LogITBackend.LogIT.web.dto.UserRequestDTO;
+import LogITBackend.LogIT.DTO.UserRequestDTO;
 import LogITBackend.LogIT.domain.Users;
 
 public interface UserCommandService {
